+++ sdk/identity/azure_identity/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 - `AzureDeveloperCliCredential` authenticates the identity logged in to the [Azure Developer CLI](https://learn.microsoft.com/azure/developer/azure-developer-cli/overview).
+- Added the `AzureDeveloperCliCredential` to the `DefaultAzureCredential`.
 
 ### Breaking Changes